chore: Доработать код для сохранения лекарств в базе и их отображения на экране
Изменил модель базы данных, так как лекарства могут быть и половинками
Добавил алерт контроллер
Добавил кнопки сохранение и добавления на экранах лекарства (кодом)
Переписал код для получения данных из базы лекарств
Немного доработал код для получения лекарств из базы

+++ MedicinesV2/Models/CoreData/Medicine+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension Medicine {
         return NSFetchRequest<Medicine>(entityName: "Medicine")
     }
 
-    @NSManaged public var number: Int64
+    @NSManaged public var number: Float
     @NSManaged public var title: String?
     @NSManaged public var firstAidKit: FirstAidKit?
